requirements.txt
fastapi
uvicorn
pydantic
python-multipart
openai-whisper
ffmpeg-python
numpy
torch